bandsync-backend/python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
librosa==0.10.1
numpy==1.26.2
werkzeug==3.0.1
scipy==1.11.4
madmom==0.16.1
openai==1.12.0
python-dotenv==1.0.0
gunicorn==21.2.0
